+++ lib/src/email/email_repository.dart
@@ -9,10 +9,34 @@ import 'package:logging/logging.dart';
 class EmailRepository {
   final Logger _log = Logger('EmailRepository');
 
+  static const String _pathProfile =
+      "https://graph.microsoft.com/v1.0/me/messages";
+
   static const String _pathMessages =
       "https://graph.microsoft.com/v1.0/me/messages";
+
   static const String _pathSend =
       "https://graph.microsoft.com/v1.0/me/sendMail";
+
+  Future<void> pathProfile(
+      {required HttppClient client,
+        String? accessToken,
+        String? filter,
+        void Function(HttppResponse)? onSuccess,
+        required void Function(HttppResponse) onResult,
+        void Function(Object)? onError}) {
+    String queryParams = filter != null ? "?$filter&\$count=true" : "?\$count=true";
+    HttppRequest request = HttppRequest(
+        uri: Uri.parse(_pathProfile + queryParams),
+        verb: HttppVerb.GET,
+        headers: HttppHeaders.typical(bearerToken: accessToken),
+        timeout: const Duration(seconds: 30),
+        onSuccess: onSuccess,
+        onResult: onResult,
+        onError: onError);
+    _log.finest('${request.verb.value} — ${request.uri}');
+    return client.request(request);
+  }
 
   Future<void> message(
       {required HttppClient client,