+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.10
+
+* Added countInbox method to tiki strategy microsoft
+* Add last page index parameter
+
 ## 0.1.9
 
 * Revert list unsubscribe feature until further investigation